Separating entities to avoid circular dependencies
